fix: make envelopDB connection optional

+++ src/writer/delete-email-subscription.ts
@@ -3,6 +3,10 @@ import { envelopDB } from '../helpers/mysql';
 type Message = { address: string };
 
 export async function verify(message: Message): Promise<boolean> {
+  if (!envelopDB) {
+    return Promise.reject('not supported');
+  }
+
   const result = await envelopDB.queryAsync(
     'SELECT * FROM subscribers WHERE address = ? AND verified > 0 LIMIT 1',
     [message.address]
@@ -16,6 +20,10 @@ export async function verify(message: Message): Promise<boolean> {
 }
 
 export async function action(message: Message): Promise<void> {
+  if (!envelopDB) {
+    return Promise.reject('not supported');
+  }
+
   await envelopDB.queryAsync('DELETE FROM subscribers WHERE address = ? AND verified > 0 LIMIT 1', [
     message.address
   ]);
